+++ tests/Handler/CurlMultiHandlerTest.php
@@ -1302,6 +1302,25 @@ class CurlMultiHandlerTest extends TestCase
         self::assertHandlerShareWasCreated();
     }
 
+    public function testTransportSharingPassesShareStateToFactory(): void
+    {
+        self::skipIfCurlShareIsUnavailable();
+
+        $handler = new CurlMultiHandler([
+            'transport_sharing' => TransportSharing::HANDLER_PREFER,
+        ]);
+
+        $factory = \Closure::bind(static function (CurlMultiHandler $handler) {
+            return $handler->factory;
+        }, null, CurlMultiHandler::class)($handler);
+
+        $opaque = \Closure::bind(static function (CurlFactory $factory): bool {
+            return $factory->opaqueShareConnectionCache;
+        }, null, CurlFactory::class)($factory);
+
+        self::assertFalse($opaque);
+    }
+
     public function testPersistentPreferTransportSharingOptionAppliesCurlShare(): void
     {
         self::skipIfCurlShareIsUnavailable();